feat: add PortfolioItem
This commit updates the App component to use the PortfolioItem component instead of the Accordion and Tag components. The Accordion and Tag components are no longer needed and have been removed. This change improves the structure and organization of the code, making it easier to manage and maintain. Additionally, the PortfolioItem component provides better functionality and flexibility for displaying portfolio items.